# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FIN-SENSE API
  description: >
    Vietnamese financial news sentiment API.
    Measures news-coverage tone only — not a buy/sell/hold signal.
  version: "1.0.0"

servers:
  - url: https://finsense-api.onrender.com/api
    description: Production (Render)
  - url: http://localhost:8000/api
    description: Local development

security: []

tags:
  - name: dashboard
    description: Public dashboard endpoints
  - name: ticker
    description: Ticker detail endpoints
  - name: audit
    description: Admin audit panel endpoints (JWT required)
  - name: auth
    description: Authentication

paths:

  # ─────────────────────────────────────────
  # DASHBOARD
  # ─────────────────────────────────────────

  /dashboard/overview:
    get:
      tags: [dashboard]
      summary: Dashboard metric cards + Fear & Greed gauge
      description: >
        Returns article count, event count, last updated timestamp,
        and gauge data for the selected time window.
        Tickers tracked is a frontend constant — not returned here.
      parameters:
        - $ref: '#/components/parameters/window'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardOverview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'

  /dashboard/feed:
    get:
      tags: [dashboard]
      summary: Top tickers or top events feed
      description: >
        Returns top 5 items per page.
        sort=ticker → ranked by event volume, includes S_final score.
        sort=event → ranked by article volume, no sentiment score.
        Use page param for load more.
      parameters:
        - $ref: '#/components/parameters/window'
        - name: sort
          in: query
          required: true
          schema:
            type: string
            enum: [ticker, event]
          description: Feed mode — top tickers by event volume or top events by article volume
        - $ref: '#/components/parameters/page'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardFeed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'

  # ─────────────────────────────────────────
  # TICKER DETAIL
  # ─────────────────────────────────────────

  /ticker/{symbol}:
    get:
      tags: [ticker]
      summary: Ticker identity card + live sentiment gauge
      description: >
        Returns company info and live S_final score for the selected window.
        Score is computed at request time from event_clusters — not from daily_sentiment_history.
      parameters:
        - $ref: '#/components/parameters/symbol'
        - $ref: '#/components/parameters/window'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Invalid window value (must be one of 24h/48h/72h)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /ticker/{symbol}/history:
    get:
      tags: [ticker]
      summary: EOD sentiment + price history for dual-axis chart
      description: >
        Returns daily sentiment scores and closing prices from daily_sentiment_history.
        Used for the 7d/30d/90d chart on the ticker detail page.
      parameters:
        - $ref: '#/components/parameters/symbol'
        - name: days
          in: query
          required: false
          schema:
            type: integer
            enum: [7, 30, 90]
            default: 30
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerHistory'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Invalid days value (must be one of 7/30/90)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /ticker/{symbol}/events:
    get:
      tags: [ticker]
      summary: Recent events mentioning this ticker
      description: >
        Returns a paginated list of events that mention the ticker, newest
        first — the ticker's full event history, NOT scoped to the
        24h/48h/72h window (that window only applies to the live S_final
        gauge on GET /ticker/{symbol}). Page 2, 3, ... keeps going back as
        far as event_clusters has data for this ticker.
      parameters:
        - $ref: '#/components/parameters/symbol'
        - $ref: '#/components/parameters/page'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerEvents'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: page < 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  # ─────────────────────────────────────────
  # AUTH
  # ─────────────────────────────────────────

  /auth/login:
    post:
      tags: [auth]
      summary: Admin login — returns JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: bearer
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─────────────────────────────────────────
  # AUDIT
  # ─────────────────────────────────────────

  /audit/events:
    get:
      tags: [audit]
      summary: Audit event queue — unaudited first
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditQueue'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

  /audit/events/{cluster_id}/{source}:
    patch:
      tags: [audit]
      summary: Approve or correct a source entry within an event
      description: >
        Approves or corrects the AI extraction for one source within a cluster.
        On correct — updates scores directly on event_clusters source_breakdown,
        flips is_audited to true, writes immutable record to audit_log.
        On approve — flips is_audited to true, writes to audit_log with no score changes.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/cluster_id'
        - name: source
          in: path
          required: true
          schema:
            type: string
            enum: [CafeF, VnExpress]
          description: >
            Matches backend.core.config.pipeline_settings.RSS_FEEDS — update this
            enum if a source is added or removed there.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditAction'
      responses:
        '200':
          description: Audit action recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cluster_id:
                    type: string
                  source:
                    type: string
                  action_type:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  /audit/log:
    get:
      tags: [audit]
      summary: Immutable audit history log
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

# ─────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    window:
      name: window
      in: query
      required: false
      schema:
        type: string
        enum: [24h, 48h, 72h]
        default: 24h
      description: Rolling time window for live sentiment computation

    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for paginated results (5 items per page)

    symbol:
      name: symbol
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Ticker'
      example: HPG
      description: Ticker symbol (uppercase)

    cluster_id:
      name: cluster_id
      in: path
      required: true
      schema:
        type: string
        example: evt_hpg_q2_2026

  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Action not permitted (e.g. attempt to write to frozen_test_set)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:

    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string

    SentimentScore:
      type: number
      format: float
      minimum: -1.0
      maximum: 1.0
      nullable: true
      description: Continuous sentiment score. null means no valid events in window.

    # Frozen vocabularies — kept in sync by hand with backend/core/enums.py.
    # These are closed sets on purpose: historical sentiment records are labeled
    # against this exact vocabulary. Do not add/remove values here without a
    # corresponding backend enum change (and re-running the frozen test set).

    Ticker:
      type: string
      description: VN30 ticker symbol. Mirrors backend/core/enums.py::Ticker.
      enum:
        - ACB
        - BID
        - BSR
        - CTG
        - FPT
        - GAS
        - GVR
        - HDB
        - HPG
        - LPB
        - MBB
        - MCH
        - MSN
        - MWG
        - SAB
        - SHB
        - SSB
        - SSI
        - STB
        - TCB
        - TCX
        - VCB
        - VHM
        - VIB
        - VIC
        - VJC
        - VNM
        - VPB
        - VPL
        - VRE

    Concept:
      type: string
      description: Macro/sector concept taxonomy. Mirrors backend/core/enums.py::Concept.
      enum:
        - BANKING
        - CONSUMER_DISCRETIONARY
        - CONSUMER_STAPLES
        - ENERGY
        - INDUSTRIALS
        - MACRO
        - MATERIALS
        - REAL_ESTATE
        - SECURITIES
        - TECHNOLOGY

    # ── Dashboard ──

    DashboardOverview:
      type: object
      properties:
        window:
          type: string
          example: 24h
        last_updated:
          type: string
          format: date-time
        articles_count:
          type: integer
        events_count:
          type: integer
        gauge:
          type: object
          properties:
            score:
              $ref: '#/components/schemas/SentimentScore'
            positive_count:
              type: integer
            neutral_count:
              type: integer
            negative_count:
              type: integer

    DashboardFeed:
      type: object
      properties:
        sort:
          type: string
          enum: [ticker, event]
        window:
          type: string
        page:
          type: integer
        has_more:
          type: boolean
        items:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/TickerFeedItem'
              - $ref: '#/components/schemas/EventFeedItem'

    TickerFeedItem:
      type: object
      properties:
        rank:
          type: integer
        ticker:
          $ref: '#/components/schemas/Ticker'
        company_name:
          type: string
          example: Hoa Phat Group
          description: >
            TODO: no ticker → company-name mapping currently exists in the
            backend (backend/core/enums.py only defines symbols). Source this
            before implementing the endpoint, or drop the field.
        event_count:
          type: integer
        sentiment_score:
          $ref: '#/components/schemas/SentimentScore'

    EventFeedItem:
      type: object
      properties:
        rank:
          type: integer
        cluster_id:
          type: string
        event_title:
          type: string
        article_count:
          type: integer
        sources:
          type: object
          description: >
            Keyed by source name. Currently CafeF and VnExpress (see
            backend/core/config.py::PipelineSettings.RSS_FEEDS) — add a source
            here if a new one is registered in the scraper.
          properties:
            CafeF:
              type: integer
            VnExpress:
              type: integer
        tickers:
          type: array
          items:
            $ref: '#/components/schemas/Ticker'

    # ── Ticker Detail ──

    TickerDetail:
      type: object
      properties:
        ticker:
          $ref: '#/components/schemas/Ticker'
        company_name:
          type: string
          description: >
            Sourced from backend/core/ticker_metadata.py (FS-39) — resolves the
            TODO that used to sit here (see TickerFeedItem.company_name, still
            open for the dashboard feed).
        sector:
          $ref: '#/components/schemas/Concept'
          description: Primary sector, from ticker_metadata.py (FS-39)
        window:
          type: string
        sentiment_score:
          $ref: '#/components/schemas/SentimentScore'
        gauge:
          type: object
          properties:
            score:
              $ref: '#/components/schemas/SentimentScore'
            positive_count:
              type: integer
            neutral_count:
              type: integer
            negative_count:
              type: integer
        article_count:
          type: integer
          description: Total event_coverage.total_articles across events mentioning this ticker in the window
        event_count:
          type: integer
          description: Count of events mentioning this ticker in the window
        last_updated:
          type: string
          format: date-time
          nullable: true
          description: Most recent updated_at among events mentioning this ticker — null if none
        is_empty_state:
          type: boolean
          description: true if no valid events in window — UI shows neutral grey gauge

    TickerHistory:
      type: object
      properties:
        ticker:
          $ref: '#/components/schemas/Ticker'
        days:
          type: integer
          description: The requested range (7/30/90) — NOT the number of rows actually returned. See data.
        data:
          type: array
          description: >
            Oldest -> newest, one row per calendar day that exists in
            daily_sentiment_history. Can be shorter than `days` if the EOD
            batch hasn't run that many times yet — not an error. Nulls in
            either field pass through untouched: no interpolation, no gap-filling.
          items:
            type: object
            properties:
              date:
                type: string
                format: date
                example: "2026-06-27"
              daily_sentiment_score:
                $ref: '#/components/schemas/SentimentScore'
              closing_price:
                type: number
                nullable: true

    TickerEvents:
      type: object
      properties:
        ticker:
          $ref: '#/components/schemas/Ticker'
        page:
          type: integer
        has_more:
          type: boolean
        items:
          type: array
          items:
            type: object
            properties:
              cluster_id:
                type: string
              event_title:
                type: string
              created_at:
                type: string
                format: date-time
              article_count:
                type: integer
                description: event_coverage.total_articles — articles ingested, higher than len(source_breakdown) by design
              sentiment_score:
                $ref: '#/components/schemas/SentimentScore'
                description: This ticker's entry from aggregated_analysis.ticker_sentiments — null if no source cleared AI_CONFIDENCE_THRESHOLD
              source_breakdown:
                type: array
                description: >
                  One row per source (not per article — each source contributes
                  exactly one representative ai_response). Sources below
                  AI_CONFIDENCE_THRESHOLD are excluded entirely, not shown with a
                  null/greyed score.
                items:
                  type: object
                  properties:
                    source:
                      type: string
                      example: CafeF
                    score:
                      type: number
                      format: float
                      minimum: -1.0
                      maximum: 1.0
                      description: This source's own score for this ticker, from source_breakdown[].ai_response.ticker_sentiments[]
                    article_title:
                      type: string
                      description: >
                        TODO: representative_article does not persist its own
                        title yet (see backend/core/schemas/event_cluster.py::
                        RepresentativeArticle) — falls back to the cluster's
                        event_title until the pipeline stores one per article.
                    article_url:
                      type: string

    # ── Audit ──

    AuditAction:
      type: object
      required: [action_type]
      properties:
        action_type:
          type: string
          enum: [approve, correct]
        error_type:
          type: string
          nullable: true
          description: Required when action_type is correct
          enum:
            - No error
            - Wrong magnitude
            - Wrong direction
            - Wrong ticker
            - Missed ticker
        corrected_ticker_sentiments:
          type: array
          nullable: true
          description: Only include tickers being corrected — not the full list
          items:
            type: object
            properties:
              ticker:
                $ref: '#/components/schemas/Ticker'
              score:
                $ref: '#/components/schemas/SentimentScore'
        corrected_concept_sentiments:
          type: array
          nullable: true
          description: Only include concepts being corrected — not the full list
          items:
            type: object
            properties:
              concept:
                $ref: '#/components/schemas/Concept'
              score:
                $ref: '#/components/schemas/SentimentScore'

    AuditQueue:
      type: object
      properties:
        page:
          type: integer
        has_more:
          type: boolean
        total_pending:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              cluster_id:
                type: string
              event_title:
                type: string
              created_at:
                type: string
                format: date-time
              source_breakdown:
                type: array
                items:
                  type: object
                  properties:
                    source:
                      type: string
                    is_audited:
                      type: boolean
                    ai_confidence:
                      type: number
                    ticker_sentiments:
                      type: array
                      items:
                        type: object
                        properties:
                          ticker:
                            $ref: '#/components/schemas/Ticker'
                          score:
                            $ref: '#/components/schemas/SentimentScore'
                    concept_sentiments:
                      type: array
                      items:
                        type: object
                        properties:
                          concept:
                            $ref: '#/components/schemas/Concept'
                          score:
                            $ref: '#/components/schemas/SentimentScore'

    AuditLog:
      type: object
      properties:
        page:
          type: integer
        has_more:
          type: boolean
        items:
          type: array
          items:
            type: object
            properties:
              cluster_id:
                type: string
              source:
                type: string
              admin_id:
                type: string
              admin_name:
                type: string
              action_type:
                type: string
              error_type:
                type: string
                nullable: true
              old_ticker_sentiments:
                type: array
                items:
                  type: object
                  properties:
                    ticker:
                      $ref: '#/components/schemas/Ticker'
                    score:
                      $ref: '#/components/schemas/SentimentScore'
              new_ticker_sentiments:
                type: array
                nullable: true
                items:
                  type: object
                  properties:
                    ticker:
                      $ref: '#/components/schemas/Ticker'
                    score:
                      $ref: '#/components/schemas/SentimentScore'
              old_concept_sentiments:
                type: array
                items:
                  type: object
                  properties:
                    concept:
                      $ref: '#/components/schemas/Concept'
                    score:
                      $ref: '#/components/schemas/SentimentScore'
              new_concept_sentiments:
                type: array
                nullable: true
                items:
                  type: object
                  properties:
                    concept:
                      $ref: '#/components/schemas/Concept'
                    score:
                      $ref: '#/components/schemas/SentimentScore'
              performed_at:
                type: string
                format: date-time